requirements.txt
python-telegram-bot==12.4.2
pymongo==3.10.1